@@ -1,30 +1,28 @@
 {
-  "name": "SearStars",
+  "name": "sear-stars",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "build": "vite build",
-    "lint": "npx eslint \"src/**/*.{ts,vue}\"",
-    "dev": "vite --port 8124",
-    "lint:fix": "npm run lint -- --fix"
+    "dev": "vite",
+    "build": "vue-tsc && vite-ssg build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "luxon": "^1.25.0",
-    "vue": "^3.0.0-rc.10",
-    "vue-router": "^4.0.0-beta.9"
+    "luxon": "^3.3.0",
+    "vue": "^3.2.47",
+    "vue-router": "^4.2.2"
   },
   "devDependencies": {
-    "@types/luxon": "^1.24.4",
-    "@typescript-eslint/eslint-plugin": "^4.1.1-alpha.7",
-    "@typescript-eslint/parser": "^4.1.1-alpha.7",
-    "@vue/compiler-sfc": "^3.0.0-rc.10",
-    "eslint": "^7.8.1",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-vue": "^7.0.0-beta.3",
-    "sass": "^1.26.10",
-    "typescript": "^4.0.2",
-    "vite": "^1.0.0-rc.4",
-    "vue-cli-plugin-tailwind": "~1.4.2"
+    "@types/luxon": "^3.3.0",
+    "@vitejs/plugin-vue": "^4.1.0",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.24",
+    "prettier": "^2.8.8",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9",
+    "vite-ssg": "^0.22.2",
+    "vue-tsc": "^1.4.2"
   }
 }
