@@ -3,12 +3,12 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue'
+import { defineComponent } from 'vue';
 
 export default defineComponent({ name: 'Loader' })
 </script>
 
-<style lang="scss" scoped>
+<style scoped>
 .loader,
 .loader:before,
 .loader:after {
